alexa/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
requests==2.26.0
aws-psycopg2==1.2.1
